+++ effectful-core/CHANGELOG.md
@@ -18,6 +18,7 @@
 * Add a `SeqForkUnlift` strategy to support running unlifting functions outside
   of the scope of effects they capture.
 * Ensure that a `LocalEnv` is only used in a thread it belongs to.
+* Add the `Effectful.Exception` module with appropriate re-exports.
 * **Breaking changes**:
   - `localSeqLend`, `localLend`, `localSeqBorrow` and `localBorrow` now take a
     list of effects instead of a single one.